Rename and reuse original dialogs, drop popover
