--- src/panels/config/devices/device-detail/show-dialog-device-automation.ts
@@ -1,24 +1,0 @@
-// When new_triggers_conditions labs feature is promoted, this file can be removed.
-import { fireEvent } from "../../../../common/dom/fire_event";
-import type { DeviceRegistryEntry } from "../../../../data/device/device_registry";
-import type { EntityRegistryEntry } from "../../../../data/entity/entity_registry";
-
-export interface DeviceAutomationDialogParams {
-  device: DeviceRegistryEntry;
-  entityReg: EntityRegistryEntry[];
-  script?: boolean;
-}
-
-export const loadDeviceAutomationDialog = () =>
-  import("./ha-device-automation-dialog");
-
-export const showDeviceAutomationDialog = (
-  element: HTMLElement,
-  detailParams: DeviceAutomationDialogParams
-): void => {
-  fireEvent(element, "show-dialog", {
-    dialogTag: "dialog-device-automation",
-    dialogImport: loadDeviceAutomationDialog,
-    dialogParams: detailParams,
-  });
-};
